requirements.txt
genanki
pandas
litellm
python-dotenv
